requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
black==19.10b0
blynklib==0.2.6
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
docopt==0.6.2
idna==2.9
Js2Py==0.70
mpris2==1.0.2
multidict==4.7.6
packaging==20.4
pathspec==0.8.0
pipwin==0.5.0
pkg-resources==0.0.0
playerdo==2.0
PyAudio==0.2.11
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.2
pySmartDL==1.3.3
python-dotenv==0.13.0
pyttsx3==2.87
pytz==2020.1
regex==2020.6.8
requests==2.23.0
six==1.15.0
slack==0.0.2
slackclient==2.7.1
soupsieve==2.0.1
SpeechRecognition==3.8.1
toml==0.10.1
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.9
wikipedia==1.4.0
yarl==1.4.2
